Move useEventClusterData() to separate module
